fixed settings only loaded once on youtube
Case: youtube uses a hack where the user clicks videos, urls changes but
it is the same window object. i.e settings are loaded once
similar to other react apps or backbone apps using url hashtag for
history

this detects that and sends an event so the custom config can react to
it instead of being executed once on load.