fix youtube bw not toggling
when adding via mobile a youtube link, it adds extra parameters (crap).
when creating a youtube playlist of bookmarked youtube videos with tylw,
the crap is lost and therefore no match is found when attempting to
toggle with bw.

This solution adds API functions to match bookmarks by anything that
starts with a string X.

Usually, video id is the first parameter and shouldn't created a problem